third_party/blink/web_tests/platform/generic/svg/dom/SVGScriptElement/script-set-href-expected.txt

Parser-created script elements:

Already containing text: PASS

Already specifying missing source: PASS

Already specifying valid source: PASS

No text and no source: PASS

Self-closing <script> tag: PASS

Using setAttribute: PASS

Dynamically-inserted script elements:

Set before insertion: PASS

Set after insertion: PASS

Set twice before and once after insertion: PASS
